move add_shots
